+++ src/Prod/Db/AbstractTable.php
@@ -15,7 +15,7 @@ use Drupal\Prod\Stats\Task;
 abstract class AbstractTable extends ProdObject implements TableInterface, StatsProviderInterface
 {
 
-    
+
     /**
      * Analysed database name
      *
@@ -74,14 +74,14 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
     private $user_group;
 
     /**
-     * record id in prod_db_stats 
+     * record id in prod_db_stats
      *
      * @var int
      */
     private $id;
 
     /**
-     * record timestamp 
+     * record timestamp
      *
      * @var int
      */
@@ -105,7 +105,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * @param string $db_name The table database
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function __construct($identifier, $db_name)
@@ -122,7 +122,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
         return $this;
     }
 
-    public function getDbName() 
+    public function getDbName()
     {
         if ( !isset($this->db_name)) {
             throw new DevelopperException('db_name is not set');
@@ -137,7 +137,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
         return $this;
     }
 
-    public function getDbIdentifier() 
+    public function getDbIdentifier()
     {
         if ( !isset($this->db_identifier)) {
             throw new DevelopperException('db_identifier is not set yet');
@@ -146,7 +146,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
     }
 
     /**
-     * Set the table name 
+     * Set the table name
      *
      * @param string $name The name
      *
@@ -162,7 +162,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * Get the table name
      *
      * @return string
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getTable()
@@ -172,14 +172,14 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
         }
         return $this->table;
     }
-    
+
     /**
-     * Set the table size and updates the total size 
+     * Set the table size and updates the total size
      *
      * @param int $size The size
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function setSize($size)
@@ -193,7 +193,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * Get the table size
      *
      * @return int
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getSize()
@@ -203,14 +203,14 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
         }
         return $this->size;
     }
-    
+
     /**
      * Set the table index size and updates the total size
      *
      * @param int $size The size
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function setIndexSize($size)
@@ -224,7 +224,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * Get the table size
      *
      * @return int
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getIndexSize()
@@ -234,12 +234,12 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
         }
         return $this->index_size;
     }
-    
+
     /**
      * Get the table+index size
      *
      * @return int
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getTotalSize()
@@ -256,7 +256,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * @param int $rows The number of rows
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function setRows($rows)
@@ -269,7 +269,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * Get the number of rows for this table
      *
      * @return int
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getRows()
@@ -279,14 +279,14 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
         }
         return $this->rows;
     }
-    
+
     /**
      * Set the user (human) group for this table
      *
      * @param int $group The group
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function setUserGroup($group)
@@ -299,7 +299,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * Get the user (human) group for this table
      *
      * @return string or NULL
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getUserGroup()
@@ -315,7 +315,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * @param boolean $bool The flag value
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function flagIsDatabase($bool)
@@ -326,11 +326,11 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
 
     /**
      * Get the is_database flag,
-     * TRUE if this table is in fact not a "table" but 
+     * TRUE if this table is in fact not a "table" but
      * instead an aggregate of tables.
      *
      * @return boolean
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getIsDatabase()
@@ -347,7 +347,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * @param int $rows The number of rows to add
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function addRows($rows)
@@ -365,7 +365,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * @param int $size The size
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function addSize($size)
@@ -384,7 +384,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * @param int $size The size
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function addIndexSize($size)
@@ -396,14 +396,14 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
         $this->total_size = (isset($this->size))? $this->index_size + $this->size : $this->index_size;
         return $this;
     }
-    
+
     /**
      * Save (upsert) the table record in prod_db_stats table.
      * This function is also responsible for feeding the
      * object with the database id (setId())
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function save()
@@ -452,7 +452,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
         } catch (Exception $e) {
             throw new DbAnalyzerException(__METHOD__ . ": Unable to reload the table record. " . $e->getMessage());
         }
-        
+
     }
 
     /**
@@ -461,7 +461,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * @param int $id The id
      *
      * @return TableInterface
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function setId($id)
@@ -474,7 +474,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * Get the table internal id
      *
      * @return int
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getId()
@@ -486,7 +486,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
     }
 
     /**
-     * Set the record timestamp 
+     * Set the record timestamp
      *
      * @param int $timestamp The UNIX timestamp
      *
@@ -502,7 +502,7 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * Get the record timestamp
      *
      * @return int an UNIX timestamp
-     * 
+     *
      * @throws Drupal\Prod\Error\DbAnalyzerException
      */
     public function getTimestamp()
@@ -519,12 +519,12 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      * @return array List of StatInterface, the array key
      *               is the stat_col
      */
-    public function getStatsList() 
+    public function getStatsList()
     {
         if (!isset($this->id)) {
             throw new DevelopperException('Cannot create stats elements before having a real record id, maybe save this object before?');
         }
-        
+
         $stat_size = new Stat(
             $this->id,
             'size',
@@ -581,41 +581,41 @@ abstract class AbstractTable extends ProdObject implements TableInterface, Stats
      *
      * @return array keyed by setting name
      */
-    public function getDefaultRrdSettings($stat_col) 
+    public function getDefaultRrdSettings($stat_col)
     {
         return self::getDefaultRrdSettingsDb();
     }
 
-    public static function getDefaultRrdSettingsDb() 
+    public static function getDefaultRrdSettingsDb()
     {
         $defaults =& drupal_static(__METHOD__);
 
         if (!isset($defaults)) {
             // get generic defaults
             $defaults = array(
-                    'interval' => variable_get(
+                    'interval' => (int) variable_get(
                         'prod_default_rrd_interval',
                         300
                     ),
-                    'points_per_graph' => variable_get(
+                    'points_per_graph' => (int) variable_get(
                         'prod_default_rrd_points_per_graph',
                         300
                     ),
-                    'points_per_aggregate' => variable_get(
+                    'points_per_aggregate' => (int) variable_get(
                         'prod_default_rrd_points_per_aggregate',
                         5
                     ),
             );
             // check for db overrides
-            $defaults['interval'] = variable_get(
+            $defaults['interval'] = (int) variable_get(
                   'prod_default_rrd_interval_db',
                    $defaults['interval']
             );
-            $defaults['points_per_graph'] = variable_get(
+            $defaults['points_per_graph'] = (int) variable_get(
                   'prod_default_rrd_points_per_graph_db',
                    $defaults['points_per_graph']
             );
-            $defaults['points_per_aggregate'] = variable_get(
+            $defaults['points_per_aggregate'] = (int) variable_get(
                   'prod_default_rrd_points_per_aggregate_db',
                    $defaults['points_per_aggregate']
             );
